requirements.txt
streamlit==1.20.0
sentence-transformers==2.2.2
numpy==1.24.4
tqdm
google-generativeai
faiss-cpu==1.7.4.post2
pyngrok
